requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.13.0
async-timeout==5.0.1
attrs==26.1.0
azure-ai-contentsafety==1.0.0
azure-ai-projects==2.0.1
azure-common==1.1.28
azure-core==1.39.0
azure-cosmos==4.16.0b2
azure-identity==1.25.3
azure-search-documents==11.6.0
azure-storage-blob==12.28.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
colorama==0.4.6
cryptography==46.0.7
distro==1.9.0
exceptiongroup==1.3.1
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
isodate==0.7.2
jiter==0.14.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.15
langchain-azure-ai==1.2.1
langchain-core==1.2.29
langchain-openai==1.1.12
langchainhub==0.1.21
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.31
msal==1.36.0
msal-extensions==1.3.1
multidict==6.7.1
numpy==2.2.6
openai==2.31.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.1
propcache==0.4.1
pycparser==3.0
pydantic==2.13.0
pydantic_core==2.46.0
PyJWT==2.12.1
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
tenacity==9.1.4
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
